Object loading progress
Need to move the loaded data to a holding structure before transfer over to OpenGL
